+++ test/server-sql/models/index.ts
@@ -1,4 +1,4 @@
 // See `test/server-mongo/models/index.ts`'s identical doc comment - the SQL twin.
-export { MailboxSQL } from "@rapidmx/restapi/sql";
+export { CalendarEventAttendeeLinkSQL, MailboxSQL } from "@rapidmx/restapi/sql";
 export { VideoMeetingSQL } from "../../../src/models/sql/VideoMeetingSQL.js";
 export { VideoMeetingInviteeSQL } from "../../../src/models/sql/VideoMeetingInviteeSQL.js";